refactor: controller 이름 변경 및 패키지화
