+++ src/main/java/blackjack/Application.java
@@ -1,15 +1,17 @@
 package blackjack;
 
+import blackjack.controller.BlackjackController;
 import blackjack.view.InputView;
 import blackjack.view.OutputView;
 
 class Application {
 
     public static void main(String[] args) {
-        InputView inputView = new InputView();
-        OutputView outputView = new OutputView();
-        Controller controller = new Controller(inputView, outputView);
+        BlackjackController blackjackController = new BlackjackController(
+                new InputView(),
+                new OutputView()
+        );
 
-        controller.run();
+        blackjackController.run();
     }
 }
